+++ internal/e2e/rc_append_return_pos_test.go
@@ -31,12 +31,12 @@ func appendReturnPosBumpSrc(n string) string {
     return acc;
 }
 function main(): i32 {
-    var before: i32 = __heap_bump_bytes();
+    var before: i32 = (__heap_bump_bytes() as i32);
     var acc: i32[] = [];
     var i: i32 = 0;
     while (i < ` + n + `) { acc = step(acc, i); i = i + 1; }
     if (acc.len() != ` + n + `) { return 255; }
-    return (__heap_bump_bytes() - before) / 256;
+    return ((__heap_bump_bytes() as i32) - before) / 256;
 }`
 }
