+++ internal/e2e/rc_heap_bump_generic_enum_composite_payload_test.go
@@ -28,7 +28,7 @@ import (
 func genEnumCompositeBumpSrc(n string) string {
 	return `enum Wrap[T] { W(T[]), Empty }
 function main(): i32 {
-    var before: i32 = __heap_bump_bytes();
+    var before: i32 = (__heap_bump_bytes() as i32);
     var i: i32 = 0;
     var acc: i32 = 0;
     var stem: string = "alpha";
@@ -41,7 +41,7 @@ function main(): i32 {
         i = i + 1;
     }
     if (acc < 0) { return acc; }
-    return __heap_bump_bytes() - before;
+    return (__heap_bump_bytes() as i32) - before;
 }`
 }
